+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   colours with the `IDisposable` pattern resetting the previous colour.
 - 🆕 `AddRange` extension method bringing the functionality of `List<T>`'s
   method to all `ICollection<T>`.
+- 🆕 `PadRight` extension method on `IList<T>` to pad a list using either a
+  fixed value or a factory function.
 
 # 0.1.0
 - 🆕 `Disposable` class for managing `IDisposable` instances.